chore(travis): Setting Travis
-create .travis.yml to run the test on travis and pass to coveralls

[Finishes #169834721]
