+++ .travis.yml
@@ -1,0 +1,12 @@
+language: node_js
+node_js:
+  - "stable"
+cache:
+  directories:
+    - "node_modules"
+install:
+  npm install
+script:
+  npm test
+after_script:
+  npm run coveralls
