+++ package.json
@@ -18,18 +18,19 @@
 		"redux": "^4.0.4",
 		"redux-thunk": "^2.3.0",
 		"enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.15.1",
-    "jest": "^24.9.0",
-    "jest-enzyme": "^7.1.2",
-    "react-test-renderer": "^16.12.0"
+		"enzyme-adapter-react-16": "^1.15.1",
+		"jest": "^24.9.0",
+		"jest-enzyme": "^7.1.2",
+		"react-test-renderer": "^16.12.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
 		"build": "react-scripts build",
-		"test": "react-scripts test",
+		"test": "react-scripts test --watchAll=false",
+		"test:watch": "react-scripts test --watchAll=true",
 		"eject": "react-scripts eject",
 		"storybook": "start-storybook -p 9009 -s public",
-    "build-storybook": "build-storybook -s public"
+		"build-storybook": "build-storybook -s public"
 	},
 	"eslintConfig": {
 		"extends": "react-app"
@@ -55,8 +56,8 @@
 		"eslint-plugin-react": "^7.16.0",
 		"redux-devtools": "^3.5.0",
 		"@storybook/addon-actions": "^5.2.6",
-    "@storybook/addon-links": "^5.2.6",
-    "@storybook/addons": "^5.2.6",
-    "@storybook/react": "^5.2.6"
+		"@storybook/addon-links": "^5.2.6",
+		"@storybook/addons": "^5.2.6",
+		"@storybook/react": "^5.2.6"
 	}
 }
