+++ src/App.js
@@ -2,8 +2,7 @@ import React from 'react';
 import './App.scss';
 import { Provider } from 'react-redux';
 import Store from './reduxStore';
-import HomeComponent from './components/homeComponent';
-import logo from './zoom.png';
+import HomeComponent from './components/HomeComponent';
 
 function App() {
   return (
