+++ src/actions/defaultActions.js
@@ -1,10 +1,10 @@
 import { DEFAULT_ACTION } from './types';
 
-export const defaultAction = () => {
-    return dispatch => {
-        dispatch({
-            type: DEFAULT_ACTION,
-            payload: 'Redux works!'
-        });
-    };
-}; 
+const defaultAction = () => (dispatch) => {
+  dispatch({
+    type: DEFAULT_ACTION,
+    payload: 'Redux works!',
+  });
+};
+
+export default defaultAction;
