+++ src/actions/types.js
@@ -1,1 +1,3 @@
 export const DEFAULT_ACTION = 'DEFAULT_ACTION';
+
+export default DEFAULT_ACTION;
