+++ src/reducers/defaultReducer.js
@@ -1,17 +1,17 @@
 import { DEFAULT_ACTION } from '../actions/types';
 
 const initialState = {
-    text: ''
+  text: '',
 };
 
 export default (state = initialState, action) => {
-    switch (action.type) {
-        case DEFAULT_ACTION:
-            return {
-                ...state,
-                text: action.payload
-            }
-        default:
-            return state;
-    }
+  switch (action.type) {
+    case DEFAULT_ACTION:
+      return {
+        ...state,
+        text: action.payload,
+      };
+    default:
+      return state;
+  }
 };
