+++ src/reducers/index.js
@@ -2,5 +2,5 @@ import { combineReducers } from 'redux';
 import defaultReducer from './defaultReducer';
 
 export default combineReducers({
-    default: defaultReducer
+  default: defaultReducer,
 });
