+++ src/reduxStore.js
@@ -9,7 +9,7 @@ const middleware = [thunk];
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
 export default createStore(
-    rootReducer,
-    initialState,
-    composeEnhancers(applyMiddleware(...middleware)),
+  rootReducer,
+  initialState,
+  composeEnhancers(applyMiddleware(...middleware)),
 );
