+++ src/setupTests.js
@@ -1,5 +1,6 @@
 import Enzyme from 'enzyme';
 import EnzymeAdapter from 'enzyme-adapter-react-16';
+
 Enzyme.configure({
-    adapter: new EnzymeAdapter()
+  adapter: new EnzymeAdapter(),
 });
